Added a way to keep track of what happens as each layer is  encapsulated.
Every Layer*FlowInfo gained a new field `stats`. This field tracks two things, what was attempted, and what the outcome of that attempt was. Right now it is a simple enum of Success, or Failure(Error). (This feels very un-rust like. It is essentially the reverse of an Option type, though having a successful outcome be a None variant feels weird.)

All try_froms have been removed and replaced with a simple From. Information about what happens when another protocol parsed is retained.